# faster vertical and horizontal gradients