… tests/paths/gradientLinear2.png